requirements.txt
flask==2.2.3
numpy==1.24.2
pandas==1.5.3
torch==2.0.0
transformers==4.27.4
tqdm==4.65.0
faiss-cpu==1.7.3
requests==2.28.2
scikit-learn==1.2.2